Code version 1.2.8: Fixed the expected Map type from server. Used to be <String, Integer>, but server sends a Map<String, Long>. Tested wifi code and confirmed working.

+++ Competition/src/ca/mcgill/ecse211/team4/robot/Robot.java
@@ -93,7 +93,7 @@ public class Robot {
 	/**
 	 * Map containing the game parameters.
 	 */
-	private static Map<String, Integer> gameParameters;
+	private static Map<String, Long> gameParameters;
 	
 	/**
 	 * Can either be "Green" or "Red". This will affect strategies.
